Fix: Don't allow classes to inherit themselves. Make struct size argument the correct type. Check for possible nullptr when generating bitfields.
Clean up code:
* Convert casts to C++-style casts.
* Use ranges library.
* Use concepts library.
* Brace initialize where possible.
* Remove some unused code/assert comments.

+++ include/sdkgenny/struct.hpp
@@ -42,7 +42,7 @@ public:
     Struct* parent(Struct* parent);
 
     size_t size() const override;
-    auto size(int size) {
+    auto size(size_t size) {
         m_size = size;
         return this;
     }
